+++ core.lua
@@ -1408,7 +1408,7 @@ function RCLootCouncil:GetLootDBStatistics()
 			numTokens[entry.instance].num = entry.tierToken and numTokens[entry.instance].num + 1 or numTokens[entry.instance].num
 			count[id] = count[id] and count[id] + 1 or 1
 			responseText[id] = responseText[id] and responseText[id] or entry.response
-			if not color[id] or unpack(color[id],1,3) == unpack({1,1,1}) and #entry.color ~= 0  then -- If it's not already added
+			if (not color[id] or unpack(color[id],1,3) == unpack({1,1,1})) and (entry.color and #entry.color ~= 0)  then -- If it's not already added
 				color[id] = #entry.color ~= 0 and #entry.color == 4 and entry.color or {1,1,1}
 			end
 			if type(id) == "number" and id <= db.numMoreInfoButtons and not entry.isAwardReason and lastestAwardFound < 5 then